// frontend/src/lib/api/generated/models/ListingOrganizationEventsItem.ts
export class ListingOrganizationEventsItem {
    'id': number;
    'slug': string;
    'name': string;
    'coverImageUrl': string | null;
    'startAt': Date | null;
    'endAt': Date | null;
    'applicationStartAt': Date | null;
    'applicationEndAt': Date | null;
    'isOnline': boolean | null;
    'isOffline': boolean | null;
    'organizeCityCode': string | null;
    'organizeAddress': string | null;
    'meetingUrl'?: string | null;
    'meetingToolCode': EventMeetingToolCode | null;
    'tickets'?: Array<OrganizationEventTicketItem>;
    'totalTicketNumber': number | null;
    'status': EventStatusCode;
    'viewCount'?: number | null;
    'tags'?: Array<TagItem>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "slug",
            "baseName": "slug",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "coverImageUrl",
            "baseName": "cover_image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "startAt",
            "baseName": "start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "endAt",
            "baseName": "end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationStartAt",
            "baseName": "application_start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationEndAt",
            "baseName": "application_end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "isOnline",
            "baseName": "is_online",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isOffline",
            "baseName": "is_offline",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "organizeCityCode",
            "baseName": "organize_city_code",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizeAddress",
            "baseName": "organize_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "meetingUrl",
            "baseName": "meeting_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "meetingToolCode",
            "baseName": "meeting_tool_code",
            "type": "EventMeetingToolCode",
            "format": ""
        },
        {
            "name": "tickets",
            "baseName": "tickets",
            "type": "Array<OrganizationEventTicketItem>",
            "format": ""
        },
        {
            "name": "totalTicketNumber",
            "baseName": "total_ticket_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "EventStatusCode",
            "format": ""
        },
        {
            "name": "viewCount",
            "baseName": "view_count",
            "type": "number",
            "format": ""
        },
        {
            "name": "tags",
            "baseName": "tags",
            "type": "Array<TagItem>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingOrganizationEventsItem.attributeTypeMap;
    }

    public constructor() {
    }
}